Implement Vector.som methods as primitives
This implements Vector in the interpreter, and most of its methods as primitives.
The implementation uses the new VMVector class as representation.

The #contains: and #indexOf: can’t work it in the current SOM++ design. Because when sending a message, we yield back to the bytecode loop, which is not recursive, and thus, can’t come back to the primitive code.

The `USE_VECTOR_PRIMITIVES` compilation flag is `true` per default and allows disabling the use of the vector primitives/strategy.

The implementation supports AWFY/Vector and core-lib/Vector and loads the correct primitive methods for each implementation.

Methods are hashed with murmur3_32 to determine whether primitives are applicable.
These hashes can now be shown by running SOM++ with the -prim-hashes flag.

Adapt PrimitiveContainer to store a pair of primitives, with their bytecode hash. Also reduce the code duplication in InstallPrimitives by using template parameters and function pointers.

Co-authored-by: Stefan Marr <git@stefan-marr.de>